run db migrations in a custom resource (#1177)
### Feature or Bugfix
- Bugfix

### Detail
Currently the DB is being initialised in the DBMigrations pipeline stage
(using CodeBuild) which runs after the BackendStage.
The SavePermissions TriggerFunction runs during the deployment of
Backend, just after deployment of the DBCluster.
As a result on clean deployments the SavePermissions step will fail
because the DB is uninitialized.

To resolve that in this PR we do the following
* remove DBMigrations stage based on CodeBuild
* run DBMigrations as part of a TriggerFunction/CustomResource after the
DB deployment
* run SavePermissions TriggerFunction after the DBMigrations
TriggerFunction

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.