Enable encryption for lambda environment variables (#1225)
### Feature or Bugfix
- Bugfix

### Detail
- The environment variables for the lambda functions are not encrypted
in cdk which are identified by checkov scans. This fix is to enable kms
encryption for the lambda environment variables.

### Relates
- [<URL or Ticket>
](https://github.com/data-dot-all/dataall/issues/1201)
### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)? N/A
  - Is the input sanitized? N/A
- What precautions are you taking before deserializing the data you
consume? N/A
  - Is injection prevented by parametrizing queries? N/A
  - Have you ensured no `eval` or similar functions are used? N/A
- Does this PR introduce any functionality or component that requires
authorization? N/A
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
N/A
  - Are you logging failed auth attempts? N/A
- Are you using or adding any cryptographic features? N/A
  - Do you use a standard proven implementations? N/A
- Are the used keys controlled by the customer? Where are they stored?
the KMS keys are generated by cdk and are used to encrypt the
environment variables for all lambda functions in the lambda-api stack
- Are you introducing any new policies/roles/users? - N/A
  - Have you used the least-privilege principle? How? N/A


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.