Create RDS database snapshot before executing alembic migrations (#1267)
### Feature or Bugfix
- Feature

### Detail
Alembic migrations can get complex and in some cases we are using
alembic for not only schema migrations but also data migrations. When
moving columns with data from one table to another we might accidentally
make a mistake in a migration script. We strive to test all migration
scripts and avoid bugs in such sensitive operations, but to protect
users from the catastrophic situation in which there is a bug, a service
issue or any other exceptional situation this PR introduces the creation
of manual database snapshots before running alembic migration scripts.

This PR modifies the db_migration handler that is triggered with every
backendStack update. It checks if there are new migration scripts (if
the current head in the database is different from the new head in the
code). If True, it will create a cluster snapshot.

Remarks:
- Snapshots cannot be created when the cluster in not `available`, the
PR introduces a check to wait for this condition. If the Lambda timeout
is reached waiting for the cluster, then the CICD pipeline will fail and
will need to be retried
- During the creation of an snapshot we can still run alembic migration
scripts
- Snapshots are incremental, the first time will take a long time, but
new snapshots will be faster

### Relates
- #1258 - This PR is a good example of complex data migration
operations.

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.