Generic dataset module and specific s3_datasets module - part 5.2 (Move list queries to dataset_base or rename them) (#1282)
### Feature or Bugfix
- Refactoring

### Detail
As explained in the design for #1123 we are trying to implement a
generic `datasets_base` module that can be used by any type of datasets
in a generic way.

In this PR we:
- Restructure listDatasetsOwnedByEnvGroup as
listS3DatasetsOwnedByEnvGroup and move it into Worksheets in FE: the
reason why it is moved to Worksheets is that it is the only place where
it is used in the FE. One could argue that in the BE
listS3DatasetsOwnedByEnvGroup is part of the S3_Dataset module. The way
I see it, FE and BE are independent and their modularization strategies
fit the type of programming, what makes sense in FE might not make it in
BE. In BE queries belong to the module whose services/models they are
performing actions on, in this case s3_datasets. In FE queries belong to
the module where they are used and if a query is used by more than one
module then it can be placed in the generic `services` directory. What
is important is that we define the dependencies. In this case it is
important to make Worksheets dependent of S3_Datasets (as we do in the
index in `frontend/src/modules/Worksheets/index.js` and in
`backend/dataall/modules/worksheets/__init__.py`
- Move listDatasetsCreatedInEnvironment to datasets_base

### Relates
- #1123 
- #955 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.